+++ implementation/scipamato/scipamato-core-persistence-jooq/src/test/java/ch/difty/scipamato/core/persistence/paper/slim/JooqPaperSlimRepoIntegrationTest.java
@@ -27,7 +27,7 @@ public class JooqPaperSlimRepoIntegrationTest extends JooqTransactionalIntegrati
 
     @Test
     public void findingById_withExistingId_returnsEntity() {
-        PaperSlim paper = repo.findById((long) MAX_ID_PREPOPULATED);
+        PaperSlim paper = repo.findById(MAX_ID_PREPOPULATED);
         if (MAX_ID_PREPOPULATED > 0)
             assertThat(paper.getId()).isEqualTo(MAX_ID_PREPOPULATED);
         else
@@ -36,7 +36,7 @@ public class JooqPaperSlimRepoIntegrationTest extends JooqTransactionalIntegrati
 
     @Test
     public void findingById_withNonExistingId_returnsNull() {
-        assertThat(repo.findById(-1l)).isNull();
+        assertThat(repo.findById(-1L)).isNull();
     }
 
     @Test
